+++ Demo/Shared/GiphyAPI.swift
@@ -21,19 +21,11 @@ extension Giphy: TargetType {
             return .post
         }
     }
-    public var parameters: [String: Any]? {
-        switch self {
-        case .upload:
-            return ["api_key": "dc6zaTOxFJmzC", "username": "Moya"]
-        }
-    }
-    public var parameterEncoding: ParameterEncoding {
-        return URLEncoding.default
-    }
     public var task: Task {
         switch self {
         case let .upload(data):
-            return .upload(.multipart([MultipartFormData(provider: .data(data), name: "file", fileName: "gif.gif", mimeType: "image/gif")]))
+            let multipartFormData = [MultipartFormData(provider: .data(data), name: "file", fileName: "gif.gif", mimeType: "image/gif")]
+            return .uploadCompositeMultipart(multipartFormData, urlParameters: ["api_key": "dc6zaTOxFJmzC", "username": "Moya"])
         }
     }
     public var sampleData: Data {