+++ Single+MoyaSpec.swift
@@ -1,25 +1,21 @@
 import Quick
 import Moya
-import ReactiveSwift
+import RxSwift
 import Nimble
 
-private func signalSendingData(_ data: Data, statusCode: Int = 200) -> SignalProducer<Response, MoyaError> {
-    return SignalProducer(value: Response(statusCode: statusCode, data: data as Data, response: nil))
-}
-
-class SignalProducerMoyaSpec: QuickSpec {
+class SingleMoyaSpec: QuickSpec {
     override func spec() {
         describe("status codes filtering") {
             it("filters out unrequested status codes") {
                 let data = Data()
-                let signal = signalSendingData(data, statusCode: 10)
+                let single = Response(statusCode: 10, data: data).asSingle()
 
                 var errored = false
-                signal.filter(statusCodes: 0...9).startWithResult { event in
+                _ = single.filter(statusCodes: 0...9).subscribe { event in
                     switch event {
                     case .success(let object):
                         fail("called on non-correct status code: \(object)")
-                    case .failure:
+                    case .error:
                         errored = true
                     }
                 }
@@ -29,14 +25,14 @@ class SignalProducerMoyaSpec: QuickSpec {
 
             it("filters out non-successful status codes") {
                 let data = Data()
-                let signal = signalSendingData(data, statusCode: 404)
+                let single = Response(statusCode: 404, data: data).asSingle()
 
                 var errored = false
-                signal.filterSuccessfulStatusCodes().startWithResult { result in
-                    switch result {
+                _ = single.filterSuccessfulStatusCodes().subscribe { event in
+                    switch event {
                     case .success(let object):
                         fail("called on non-success status code: \(object)")
-                    case .failure:
+                    case .error:
                         errored = true
                     }
                 }
@@ -46,26 +42,26 @@ class SignalProducerMoyaSpec: QuickSpec {
 
             it("passes through correct status codes") {
                 let data = Data()
-                let signal = signalSendingData(data)
+                let single = Response(statusCode: 200, data: data).asSingle()
 
                 var called = false
-                signal.filterSuccessfulStatusCodes().startWithResult { _ in
+                _ = single.filterSuccessfulStatusCodes().subscribe(onSuccess: { _ in
                     called = true
-                }
+                })
 
                 expect(called).to(beTruthy())
             }
 
             it("filters out non-successful status and redirect codes") {
                 let data = Data()
-                let signal = signalSendingData(data, statusCode: 404)
+                let single = Response(statusCode: 404, data: data).asSingle()
 
                 var errored = false
-                signal.filterSuccessfulStatusAndRedirectCodes().startWithResult { result in
-                    switch result {
+                _ = single.filterSuccessfulStatusAndRedirectCodes().subscribe { event in
+                    switch event {
                     case .success(let object):
                         fail("called on non-success status code: \(object)")
-                    case .failure:
+                    case .error:
                         errored = true
                     }
                 }
@@ -75,50 +71,50 @@ class SignalProducerMoyaSpec: QuickSpec {
 
             it("passes through correct status codes") {
                 let data = Data()
-                let signal = signalSendingData(data)
+                let single = Response(statusCode: 200, data: data).asSingle()
 
                 var called = false
-                signal.filterSuccessfulStatusAndRedirectCodes().startWithResult { _ in
+                _ = single.filterSuccessfulStatusAndRedirectCodes().subscribe(onSuccess: { _ in
                     called = true
-                }
+                })
 
                 expect(called).to(beTruthy())
             }
 
             it("passes through correct redirect codes") {
                 let data = Data()
-                let signal = signalSendingData(data, statusCode: 304)
+                let single = Response(statusCode: 304, data: data).asSingle()
 
                 var called = false
-                signal.filterSuccessfulStatusAndRedirectCodes().startWithResult { _ in
+                _ = single.filterSuccessfulStatusAndRedirectCodes().subscribe(onSuccess: { _ in
                     called = true
-                }
+                })
 
                 expect(called).to(beTruthy())
             }
 
-            it("knows how to filter individual status codes") {
+            it("knows how to filter individual status code") {
                 let data = Data()
-                let signal = signalSendingData(data, statusCode: 42)
+                let single = Response(statusCode: 42, data: data).asSingle()
 
                 var called = false
-                signal.filter(statusCode: 42).startWithResult { _ in
+                _ = single.filter(statusCode: 42).subscribe(onSuccess: { _ in
                     called = true
-                }
+                })
 
                 expect(called).to(beTruthy())
             }
 
             it("filters out different individual status code") {
                 let data = Data()
-                let signal = signalSendingData(data, statusCode: 43)
+                let single = Response(statusCode: 43, data: data).asSingle()
 
                 var errored = false
-                signal.filter(statusCode: 42).startWithResult { result in
-                    switch result {
+                _ = single.filter(statusCode: 42).subscribe { event in
+                    switch event {
                     case .success(let object):
                         fail("called on non-success status code: \(object)")
-                    case .failure:
+                    case .error:
                         errored = true
                     }
                 }
@@ -130,28 +126,29 @@ class SignalProducerMoyaSpec: QuickSpec {
         describe("image maping") {
             it("maps data representing an image to an image") {
                 let image = Image.testPNGImage(named: "testImage")
-                let data = ImageJPEGRepresentation(image, 0.75)
-                let signal = signalSendingData(data!)
+                guard let data = ImageJPEGRepresentation(image, 0.75) else { fatalError("Failed creating Data from Image") }
+
+                let single = Response(statusCode: 200, data: data).asSingle()
 
                 var size: CGSize?
-                signal.mapImage().startWithResult { _ in
-                    size = image.size
-                }
+                _ = single.mapImage().subscribe(onSuccess: { image in
+                    size = image?.size
+                })
 
                 expect(size).to(equal(image.size))
             }
 
             it("ignores invalid data") {
                 let data = Data()
-                let signal = signalSendingData(data)
+                let single = Response(statusCode: 200, data: data).asSingle()
 
                 var receivedError: MoyaError?
-                signal.mapImage().startWithResult { result in
-                    switch result {
+                _ = single.mapImage().subscribe { event in
+                    switch event {
                     case .success:
                         fail("next called for invalid data")
-                    case .failure(let error):
-                        receivedError = error
+                    case .error(let error):
+                        receivedError = error as? MoyaError
                     }
                 }
 
@@ -165,15 +162,14 @@ class SignalProducerMoyaSpec: QuickSpec {
             it("maps data representing some JSON to that JSON") {
                 let json = ["name": "John Crighton", "occupation": "Astronaut"]
                 let data = try! JSONSerialization.data(withJSONObject: json, options: .prettyPrinted)
-                let signal = signalSendingData(data)
+                let single = Response(statusCode: 200, data: data).asSingle()
 
                 var receivedJSON: [String: String]?
-                signal.mapJSON().startWithResult { result in
-                    if case .success(let _json) = result,
-                        let json = _json as? [String: String] {
+                _ = single.mapJSON().subscribe(onSuccess: { json in
+                    if let json = json as? [String: String] {
                         receivedJSON = json
                     }
-                }
+                })
 
                 expect(receivedJSON?["name"]).to(equal(json["name"]))
                 expect(receivedJSON?["occupation"]).to(equal(json["occupation"]))
@@ -181,16 +177,17 @@ class SignalProducerMoyaSpec: QuickSpec {
 
             it("returns a Cocoa error domain for invalid JSON") {
                 let json = "{ \"name\": \"john }"
-                let data = json.data(using: String.Encoding.utf8)
-                let signal = signalSendingData(data!)
+                guard let data = json.data(using: .utf8) else { fatalError("Failed creating Data from JSON String") }
+
+                let single = Response(statusCode: 200, data: data).asSingle()
 
                 var receivedError: MoyaError?
-                signal.mapJSON().startWithResult { result in
-                    switch result {
+                _ = single.mapJSON().subscribe { event in
+                    switch event {
                     case .success:
                         fail("next called for invalid data")
-                    case .failure(let error):
-                        receivedError = error
+                    case .error(let error):
+                        receivedError = error as? MoyaError
                     }
                 }
 
@@ -207,13 +204,14 @@ class SignalProducerMoyaSpec: QuickSpec {
         describe("string mapping") {
             it("maps data representing a string to a string") {
                 let string = "You have the rights to the remains of a silent attorney."
-                let data = string.data(using: String.Encoding.utf8)
-                let signal = signalSendingData(data!)
+                guard let data = string.data(using: .utf8) else { fatalError("Failed creating Data from String") }
+
+                let single = Response(statusCode: 200, data: data).asSingle()
 
                 var receivedString: String?
-                signal.mapString().startWithResult { result in
-                    receivedString = result.value
-                }
+                _ = single.mapString().subscribe(onSuccess: { string in
+                    receivedString = string
+                })
 
                 expect(receivedString).to(equal(string))
             }
@@ -221,31 +219,34 @@ class SignalProducerMoyaSpec: QuickSpec {
             it("maps data representing a string at a key path to a string") {
                 let string = "You have the rights to the remains of a silent attorney."
                 let json = ["words_to_live_by": string]
-                let data = try! JSONSerialization.data(withJSONObject: json, options: .prettyPrinted)
-                let signal = signalSendingData(data)
+                guard let data = try? JSONSerialization.data(withJSONObject: json, options: .prettyPrinted) else {
+                    fatalError("Failed creating Data from JSON dictionary")
+                }
+
+                let single = Response(statusCode: 200, data: data).asSingle()
 
                 var receivedString: String?
-                signal.mapString(atKeyPath: "words_to_live_by").startWithResult { result in
-                    receivedString = result.value
-                }
+                _ = single.mapString(atKeyPath: "words_to_live_by").subscribe(onSuccess: { string in
+                    receivedString = string
+                })
 
                 expect(receivedString).to(equal(string))
             }
 
             it("ignores invalid data") {
                 let data = Data(bytes: [0x11FFFF] as [UInt32], count: 1) //Byte exceeding UTF8
-                let signal = signalSendingData(data as Data)
+                let single = Response(statusCode: 200, data: data).asSingle()
 
                 var receivedError: MoyaError?
-                signal.mapString().startWithResult { result in
-                    switch result {
+                _ = single.mapString().subscribe { event in
+                    switch event {
                     case .success:
                         fail("next called for invalid data")
-                    case .failure(let error):
-                        receivedError = error
+                    case .error(let error):
+                        receivedError = error as? MoyaError
                     }
                 }
-
+                
                 expect(receivedError).toNot(beNil())
                 let expectedError = MoyaError.stringMapping(Response(statusCode: 200, data: Data(), response: nil))
                 expect(receivedError).to(beOfSameErrorType(expectedError))