+++ Sources/Moya/MoyaProvider+Defaults.swift
@@ -7,8 +7,7 @@ public extension MoyaProvider {
             url: URL(target: target).absoluteString,
             sampleResponseClosure: { .networkResponse(200, target.sampleData) },
             method: target.method,
-            parameters: target.parameters,
-            parameterEncoding: target.parameterEncoding,
+            task: target.task,
             httpHeaderFields: target.headers
         )
     }