+++ Tests/MultiTargetSpec.swift
@@ -9,9 +9,7 @@ class MultiTargetSpec: QuickSpec {
                 let baseURL = URL(string: "http://example.com")!
                 let path = "/endpoint"
                 let method = Moya.Method.get
-                let parameters: [String: Any]? = ["key": "value"]
-                let parameterEncoding: Moya.ParameterEncoding = JSONEncoding.default
-                let task = Task.request
+                let task = Task.requestParameters(parameters: ["key": "value"], encoding: JSONEncoding.default)
                 let sampleData = "sample data".data(using: .utf8)!
                 let validate = true
                 let headers: [String: String]? = ["headerKey": "headerValue"]
@@ -32,12 +30,20 @@ class MultiTargetSpec: QuickSpec {
             }
 
             it("uses correct parameters") {
-                expect(target.parameters?["key"] as? String) == "value"
-                expect(target.parameters?.count) == 1
+                if case let .requestParameters(parameters: parameters, encoding: _) = target.task {
+                    expect(parameters["key"] as? String) == "value"
+                    expect(parameters.count) == 1
+                } else {
+                    fail("expected task type `.requestParameters`, was \(String(describing: target.task))")
+                }
             }
 
             it("uses correct parameter encoding.") {
-                expect(target.parameterEncoding is JSONEncoding) == true
+                if case let .requestParameters(parameters: _, encoding: parameterEncoding) = target.task {
+                    expect(parameterEncoding is JSONEncoding) == true
+                } else {
+                    fail("expected task type `.requestParameters`, was \(String(describing: target.task))")
+                }
             }
 
             it("uses correct method") {
@@ -45,7 +51,7 @@ class MultiTargetSpec: QuickSpec {
             }
 
             it("uses correct task") {
-                expect(String(describing: target.task)) == "request" // Hack to avoid implementing Equatable for Task
+                expect(String(describing: target.task)).to(beginWith("requestParameters")) // Hack to avoid implementing Equatable for Task
             }
 
             it("uses correct sample data") {