+++ Tests/Observable+MoyaSpec.swift
@@ -3,40 +3,12 @@ import Moya
 import RxSwift
 import Nimble
 
-#if os(iOS) || os(watchOS) || os(tvOS)
-    private func ImageJPEGRepresentation(_ image: ImageType, _ compression: CGFloat) -> Data? {
-        return UIImageJPEGRepresentation(image, compression)
-    }
-#elseif os(OSX)
-    private func ImageJPEGRepresentation(_ image: ImageType, _ compression: CGFloat) -> Data? {
-        var imageRect = CGRect(x: 0, y: 0, width: image.size.width, height: image.size.height)
-        let imageRep = NSBitmapImageRep(cgImage: image.cgImage(forProposedRect: &imageRect, context: nil, hints: nil)!)
-        return imageRep.representation(using: .JPEG, properties:[:])
-    }
-#endif
-
-// Necessary since Image(named:) doesn't work correctly in the test bundle
-private extension ImageType {
-
-    class TestClass { }
-
-    class func testPNGImage(named name: String) -> ImageType {
-        let bundle = Bundle(for: type(of: TestClass()))
-        let path = bundle.path(forResource: name, ofType: "png")
-        return Image(contentsOfFile: path!)!
-    }
-}
-
-private func observableSendingData(_ data: Data, statusCode: Int = 200) -> Observable<Response> {
-    return Observable.just(Response(statusCode: statusCode, data: data, response: nil))
-}
-
 class ObservableMoyaSpec: QuickSpec {
     override func spec() {
         describe("status codes filtering") {
             it("filters out unrequested status codes") {
                 let data = Data()
-                let observable = observableSendingData(data, statusCode: 10)
+                let observable = Response(statusCode: 10, data: data).asObservable()
 
                 var errored = false
                 _ = observable.filter(statusCodes: 0...9).subscribe { event in
@@ -55,7 +27,7 @@ class ObservableMoyaSpec: QuickSpec {
 
             it("filters out non-successful status codes") {
                 let data = Data()
-                let observable = observableSendingData(data, statusCode: 404)
+                let observable = Response(statusCode: 404, data: data).asObservable()
 
                 var errored = false
                 _ = observable.filterSuccessfulStatusCodes().subscribe { event in
@@ -74,7 +46,7 @@ class ObservableMoyaSpec: QuickSpec {
 
             it("passes through correct status codes") {
                 let data = Data()
-                let observable = observableSendingData(data)
+                let observable = Response(statusCode: 200, data: data).asObservable()
 
                 var called = false
                 _ = observable.filterSuccessfulStatusCodes().subscribe(onNext: { _ in
@@ -86,7 +58,7 @@ class ObservableMoyaSpec: QuickSpec {
 
             it("filters out non-successful status and redirect codes") {
                 let data = Data()
-                let observable = observableSendingData(data, statusCode: 404)
+                let observable = Response(statusCode: 404, data: data).asObservable()
 
                 var errored = false
                 _ = observable.filterSuccessfulStatusAndRedirectCodes().subscribe { event in
@@ -105,7 +77,7 @@ class ObservableMoyaSpec: QuickSpec {
 
             it("passes through correct status codes") {
                 let data = Data()
-                let observable = observableSendingData(data)
+                let observable = Response(statusCode: 200, data: data).asObservable()
 
                 var called = false
                 _ = observable.filterSuccessfulStatusAndRedirectCodes().subscribe(onNext: { _ in
@@ -117,7 +89,7 @@ class ObservableMoyaSpec: QuickSpec {
 
             it("passes through correct redirect codes") {
                 let data = Data()
-                let observable = observableSendingData(data, statusCode: 304)
+                let observable = Response(statusCode: 304, data: data).asObservable()
 
                 var called = false
                 _ = observable.filterSuccessfulStatusAndRedirectCodes().subscribe(onNext: { _ in
@@ -129,7 +101,7 @@ class ObservableMoyaSpec: QuickSpec {
 
             it("knows how to filter individual status code") {
                 let data = Data()
-                let observable = observableSendingData(data, statusCode: 42)
+                let observable = Response(statusCode: 42, data: data).asObservable()
 
                 var called = false
                 _ = observable.filter(statusCode: 42).subscribe(onNext: { _ in
@@ -141,7 +113,7 @@ class ObservableMoyaSpec: QuickSpec {
 
             it("filters out different individual status code") {
                 let data = Data()
-                let observable = observableSendingData(data, statusCode: 43)
+                let observable = Response(statusCode: 43, data: data).asObservable()
 
                 var errored = false
                 _ = observable.filter(statusCode: 42).subscribe { event in
@@ -162,8 +134,9 @@ class ObservableMoyaSpec: QuickSpec {
         describe("image maping") {
             it("maps data representing an image to an image") {
                 let image = Image.testPNGImage(named: "testImage")
-                let data = ImageJPEGRepresentation(image, 0.75)
-                let observable = observableSendingData(data!)
+                guard let data = ImageJPEGRepresentation(image, 0.75) else { fatalError("Failed creating Data from Image") }
+
+                let observable = Response(statusCode: 200, data: data).asObservable()
 
                 var size: CGSize?
                 _ = observable.mapImage().subscribe(onNext: { image in
@@ -175,7 +148,7 @@ class ObservableMoyaSpec: QuickSpec {
 
             it("ignores invalid data") {
                 let data = Data()
-                let observable = observableSendingData(data)
+                let observable = Response(statusCode: 200, data: data).asObservable()
 
                 var receivedError: MoyaError?
                 _ = observable.mapImage().subscribe { event in
@@ -199,7 +172,7 @@ class ObservableMoyaSpec: QuickSpec {
             it("maps data representing some JSON to that JSON") {
                 let json = ["name": "John Crighton", "occupation": "Astronaut"]
                 let data = try! JSONSerialization.data(withJSONObject: json, options: .prettyPrinted)
-                let observable = observableSendingData(data)
+                let observable = Response(statusCode: 200, data: data).asObservable()
 
                 var receivedJSON: [String: String]?
                 _ = observable.mapJSON().subscribe(onNext: { json in
@@ -214,8 +187,9 @@ class ObservableMoyaSpec: QuickSpec {
 
             it("returns a Cocoa error domain for invalid JSON") {
                 let json = "{ \"name\": \"john }"
-                let data = json.data(using: .utf8)
-                let observable = observableSendingData(data!)
+                guard let data = json.data(using: .utf8) else { fatalError("Failed creating Data from JSON String") }
+
+                let observable = Response(statusCode: 200, data: data).asObservable()
 
                 var receivedError: MoyaError?
                 _ = observable.mapJSON().subscribe { event in
@@ -242,8 +216,9 @@ class ObservableMoyaSpec: QuickSpec {
         describe("string mapping") {
             it("maps data representing a string to a string") {
                 let string = "You have the rights to the remains of a silent attorney."
-                let data = string.data(using: .utf8)
-                let observable = observableSendingData(data!)
+                guard let data = string.data(using: .utf8) else { fatalError("Failed creating Data from String") }
+
+                let observable = Response(statusCode: 200, data: data).asObservable()
 
                 var receivedString: String?
                 _ = observable.mapString().subscribe(onNext: { string in
@@ -256,8 +231,11 @@ class ObservableMoyaSpec: QuickSpec {
             it("maps data representing a string at a key path to a string") {
                 let string = "You have the rights to the remains of a silent attorney."
                 let json = ["words_to_live_by": string]
-                let data = try! JSONSerialization.data(withJSONObject: json, options: .prettyPrinted)
-                let observable = observableSendingData(data)
+                guard let data = try? JSONSerialization.data(withJSONObject: json, options: .prettyPrinted) else {
+                    fatalError("Failed creating Data from JSON dictionary")
+                }
+
+                let observable = Response(statusCode: 200, data: data).asObservable()
 
                 var receivedString: String?
                 _ = observable.mapString(atKeyPath: "words_to_live_by").subscribe(onNext: { string in
@@ -269,7 +247,7 @@ class ObservableMoyaSpec: QuickSpec {
 
             it("ignores invalid data") {
                 let data = Data(bytes: [0x11FFFF] as [UInt32], count: 1) //Byte exceeding UTF8
-                let observable = observableSendingData(data)
+                let observable = Response(statusCode: 200, data: data).asObservable()
 
                 var receivedError: MoyaError?
                 _ = observable.mapString().subscribe { event in