+++ Changelog.md
@@ -1,4 +1,9 @@
 # Next
+- Added tests for Single+Response and refactored Shared Test Helpers.
+- **Breaking Change** Replaced `parameters` & `parameterEncoding` in `TargetType` with extended `Task` cases.
+- **Breaking Change** Flattened `UploadType` and `DownloadType` into `Task` cases.
+- Added Swift 4.0 support.
+- Added all the `filter`/`map` operators that were available for `Observable<Response>` to `Single<Response>` as well.
 
 # 9.0.0-alpha.1